build:完成user rpc部分

+++ pkg/errno/errno.go
@@ -6,7 +6,7 @@ import (
 )
 
 type ErrNo struct {
-	ErrCode int64
+	ErrCode int32
 	ErrMsg  string
 }
 
@@ -14,7 +14,7 @@ func (e ErrNo) Error() string {
 	return fmt.Sprintf("err_code=%d, err_msg=%s", e.ErrCode, e.ErrMsg)
 }
 
-func NewErrNo(code int64, msg string) ErrNo {
+func NewErrNo(code int32, msg string) ErrNo {
 	return ErrNo{
 		ErrCode: code,
 		ErrMsg:  msg,
@@ -26,7 +26,7 @@ func (e ErrNo) WithMessage(msg string) ErrNo {
 	return e
 }
 
-func (e ErrNo) GetErrCode() int64 {
+func (e ErrNo) GetErrCode() int32 {
 	return e.ErrCode
 }
